+++ examples/ts/cli/components/client.ts
@@ -8,19 +8,24 @@ import {
   status,
 } from '@inworld/nodejs-sdk';
 import { ChildProcess, fork } from 'child_process';
+import * as kill from 'tree-kill';
 
 import { CLIENT_ACTION, CONVERSATION_ACTION, DISPLAY_WHEN } from './types';
+
+export interface ClientProps {
+  config?: ClientConfiguration;
+  previousState?: string;
+  text?: { displayWhen: DISPLAY_WHEN };
+  onDisconnect: () => void;
+}
 export class Client {
   private client: InworldClient;
   private config: ClientConfiguration | undefined;
   private connection: InworldConnectionService | null = null;
   private conversationProcess: ChildProcess;
+  private interactionIsEnded: boolean = false;
 
-  constructor(props: {
-    config?: ClientConfiguration;
-    text?: { displayWhen: DISPLAY_WHEN };
-    onDisconnect: () => void;
-  }) {
+  constructor(props: ClientProps) {
     this.conversationProcess = fork(`${__dirname}/conversation_process.ts`);
     this.conversationProcess.on('message', this.onConversationProcessMessage);
 
@@ -44,6 +49,12 @@ export class Client {
       })
       .setOnMessage(this.onMessage);
 
+    if (props.previousState) {
+      this.client.setSessionContinuation({
+        previousState: props.previousState,
+      });
+    }
+
     if (props.config) {
       this.config = props.config;
       this.client.setConfiguration(this.config);
@@ -51,7 +62,9 @@ export class Client {
   }
 
   closeConnection() {
-    this.conversationProcess?.kill();
+    if (this.conversationProcess?.pid) {
+      kill(this.conversationProcess.pid);
+    }
   }
 
   getConnection() {
@@ -71,7 +84,13 @@ export class Client {
     });
   }
 
+  getInteractionIsEnded() {
+    return this.interactionIsEnded;
+  }
+
   private onMessage = (packet: InworldPacket) => {
+    this.interactionIsEnded = false;
+
     // TEXT
     if (packet.isText()) {
       if (packet.routing.source.isPlayer) {
@@ -119,8 +138,17 @@ export class Client {
       });
     }
 
+    // NARRATED_ACTION
+    if (packet.isNarratedAction()) {
+      this.conversationProcess.send({
+        action: CONVERSATION_ACTION.NARRATED_ACTION,
+        packet,
+      });
+    }
+
     // INTERACTION_END
     if (packet.isInteractionEnd()) {
+      this.interactionIsEnded = true;
       this.conversationProcess.send({
         action: CONVERSATION_ACTION.END_INTERACTION,
         packet,