+++ src/entities/continuation/session_continuation.entity.ts
@@ -2,12 +2,18 @@ import { DialogPhrase, PreviousDialog } from './previous_dialog.entity';
 
 export interface SessionContinuationProps {
   previousDialog?: DialogPhrase[];
+  previousState?: string;
 }
 
 export class SessionContinuation {
+  readonly previousState: string | undefined;
   readonly previousDialog: PreviousDialog | undefined;
 
   constructor(props: SessionContinuationProps) {
-    this.previousDialog = new PreviousDialog(props.previousDialog);
+    if (props.previousDialog) {
+      this.previousDialog = new PreviousDialog(props.previousDialog);
+    }
+
+    this.previousState = props.previousState;
   }
 }